+++ app/script/App.ts
@@ -1,6 +1,7 @@
 import { Env } from './Env';
 import Pjax from '@codewithkyle/pjax';
 import DeviceManager from '@codewithkyle/device-manager';
+import { ModuleManager } from './ModuleManager';
 
 class App{
 
@@ -25,6 +26,8 @@ class App{
         // Start Device Manager
         this._deviceManager = new DeviceManager(Env.isDebug, true);
 
+        new ModuleManager();
+
         // Listen for a successful page transition event
         document.addEventListener('pjax:complete', this.handlePageLoad);
     }
@@ -34,6 +37,7 @@ class App{
      */
     private handlePageLoad:EventListener = ()=>{
         this.reinit();
+
     }
 
     /**
@@ -41,9 +45,12 @@ class App{
      * Any classes/packages that need to manage their DOM hooks should be told to reinit.
      */
     public reinit():void{
-        if(this._deviceManager !== null){
-            this._deviceManager.reinit();
-        }
+
+        // Update custom touch status class tracking
+        if(this._deviceManager !== null){ this._deviceManager.reinit(); }
+
+        // Create and destroy modules
+        ModuleManager.wrangleModules();
     }
 }