+++ public/assets/scripts/object1.05547527.js
@@ -1,0 +1,1 @@
+(window.webpackJsonp=window.webpackJsonp||[]).push([[6],{22:function(o,n){console.log("object 1")}},[[22,0]]]);