+++ public/assets/scripts/object1.19652699.js
@@ -1,1 +1,0 @@
-(window.webpackJsonp=window.webpackJsonp||[]).push([[5],{18:function(o,n){console.log("object 1")}},[[18,0]]]);